Add api, add defaults settings, experiments with human consumable data
