--- src/api.jl
@@ -1,3 +1,0 @@
-function json_api(json_input::AbstractString)::String 
-    dictionary_input = JSON.parse(json_input)
-end
